add us0002-n

+++ socialMeli/src/main/java/com/javabootcamp/socialmeli/repository/FollowRepository.java
@@ -11,6 +11,6 @@ public interface FollowRepository {
     void delete(Follow follow);
     List<User> findFollowersById(Integer id);
     List<User> findFollowedsById(Integer id);
-    Long countFollowersById(Integer id);
+    int countFollowersById(Integer id);
     Optional<Follow> findByFollowerIdAndFollowedId(Integer followerId, Integer followedId);
 }
