+++ socialMeli/src/main/java/com/javabootcamp/socialmeli/repository/FollowRepositoryImpl.java
@@ -45,8 +45,8 @@ public class FollowRepositoryImpl implements FollowRepository {
     }
 
     @Override
-    public Long countFollowersById(Integer id) {
-        return followsList
+    public int countFollowersById(Integer id) {
+        return (int) followsList
                 .stream()
                 .filter(f -> f.getFollowed().getId().equals(id))
                 .count();
